% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{add_known}
\alias{add_known}
\title{Add known counts to sims dataframe for ggplot}
\usage{
add_known(plot_df, known, start_date)
}
\arguments{
\item{plot_df}{An seiqhrf object returned from function \code{\link{seiqhrf}}.}

\item{known}{Dataframe with known compartment numbers to plot alongside
projections}

\item{start_date}{Date for day 0.}
}
\value{
dataframe with known data added
}
\description{
Add known counts to sims dataframe for ggplot
}
