% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD_arrivals.R
\name{arrivals.FUN}
\alias{arrivals.FUN}
\title{Arrivals function}
\usage{
arrivals.FUN(dat, at, seed = NULL)
}
\arguments{
\item{dat}{Input data needed.}

\item{at}{time point}

\item{seed}{random seed for checking consistency}
}
\value{
Updated dat
}
\description{
Function to handle background demographics for the SEIQHRF model.
Specifically arrivals (births and immigration). Uses the original EpiModel
code currently. A replacement that implements modelling the arrival of
infected individuals is under development -- but for now, all arrivals
go into the S compartment..
}
