% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_as.data.frame.R
\name{as.data.frame.seiqhrf}
\alias{as.data.frame.seiqhrf}
\title{Extract Model Data for Stochastic Models}
\usage{
\method{as.data.frame}{seiqhrf}(
  x,
  row.names = NULL,
  optional = FALSE,
  out = "vals",
  sim,
  qval,
  ...
)
}
\arguments{
\item{x}{An \code{EpiModel} object of class \code{icm} or \code{netsim}.}

\item{row.names}{See \code{\link{as.data.frame.default}}.}

\item{optional}{See \code{\link{as.data.frame.default}}.}

\item{out}{Data output to data frame: \code{"mean"} for row means across
simulations, \code{"sd"} for row standard deviations across simulations,
\code{"qnt"} for row quantiles at the level specified in \code{qval},
or \code{"vals"} for values from individual simulations.}

\item{sim}{If \code{out="vals"}, the simulation number to output. If not
specified, then data from all simulations will be output.}

\item{qval}{Quantile value required when \code{out="qnt"}.}

\item{...}{See \code{\link{as.data.frame.default}}.}
}
\description{
This function extracts model simulations for objects of classes
             \code{seiqhrf} into a data frame using
             the generic \code{as.data.frame} function.
}
\details{
These methods work for both \code{seiqhrf} class models. The
available output includes time-specific means, standard deviations, quantiles,
and simulation values (compartment and flow sizes) from these stochastic model
classes. Means, standard deviations, and quantiles are calculated by taking the
row summary (i.e., each row of data is corresponds to a time step) across all
simulations in the model output.
}
\examples{
## Stochastic SEIQHRF model
param <- param_seiqhrf()
init <- init_seiqhrf(s.num = 500, i.num = 1)
control <- control_seiqhrf(nsteps = 10, nsims = 3)
mod <- seiqhrf(init, control, param)

# Default output all simulation runs, default to all in stacked data.frame
as.data.frame(mod)
as.data.frame(mod, sim = 2)

# Time-specific means across simulations
as.data.frame(mod, out = "mean")

# Time-specific standard deviations across simulations
as.data.frame(mod, out = "sd")

# Time-specific quantile values across simulations
as.data.frame(mod, out = "qnt", qval = 0.25)
as.data.frame(mod, out = "qnt", qval = 0.75)

}
\keyword{extract}
