% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{ci_info_update}
\alias{ci_info_update}
\title{Internal}
\usage{
ci_info_update(obj, exp_name = "seiqhrf model")
}
\arguments{
\item{obj}{An seiqhrf object}

\item{exp_name}{Name of experiment}
}
\description{
Internal
}
