% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cum_discr_si}
\alias{cum_discr_si}
\title{Sample value from Weibull distribution}
\usage{
cum_discr_si(vecTimeSinceExp, scale, shape)
}
\arguments{
\item{vecTimeSinceExp}{?}

\item{scale}{scale parameter value}

\item{shape}{shape parameter value}
}
\value{
Value of coefficient of variation for vector
}
\description{
Function to sample values from a Weibull distribution with parameters shape
and rate
}
