% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{format_sims}
\alias{format_sims}
\title{Format seiqhrf objects into dataframe for ggplot}
\usage{
format_sims(x, time_limit, start_date)
}
\arguments{
\item{x}{An seiqhrf object returned from function \code{\link{seiqhrf}}.}

\item{time_limit}{Number of steps (days) to plot.}

\item{start_date}{Date for day 0.}
}
\value{
dataframe
}
\description{
Format seiqhrf objects into dataframe for ggplot
}
