% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{get_ci}
\alias{get_ci}
\title{Get 95\% confidence intervals}
\usage{
get_ci(x, plot_df)
}
\arguments{
\item{x}{An seiqhrf object returned from function \code{\link{seiqhrf}}.}

\item{plot_df}{Dataframe with known compartment numbers to plot alongside
projections}
}
\value{
dataframe with CIs and sd added
}
\description{
Get 95\% confidence intervals
}
