% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prev.R
\name{get_prev.FUN}
\alias{get_prev.FUN}
\title{Get previous function}
\usage{
get_prev.FUN(dat, at)
}
\arguments{
\item{dat}{Merged input parameters.}

\item{at}{Step number}
}
\value{
Updated dat
}
\description{
Utility function that collects prevalence and transition time data from each
run and stores it away in the simulation result object.
}
