% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{get_weekly_local}
\alias{get_weekly_local}
\title{Extract and plot information of local and weekly estimates from simulation}
\usage{
get_weekly_local(
  sim,
  market.share = market.share,
  icu_percent = icu_percent,
  start_date = start_date,
  show_start_date = show_start_date,
  time_limit = time_limit,
  sim_population = sim_population,
  total_population = total_population
)
}
\arguments{
\item{sim}{An \code{seiqhrf} object returned by \link{simulate_seiqhrf}.}

\item{market.share}{between 0 and 1, percentage of local hospital beds in 
the simulated unit (e.g. state)}

\item{icu_percent}{between 0 and 1, percentage of patients that should go to 
ICU among the ones that need hospitalization}

\item{start_date}{Epidemic start date. Default is 'na', if not provided will 
plot week numbers, if provided will plot the first day (Sunday) of the
week.}

\item{show_start_date}{First date to show in plots. Use ymd format. If FALSE,
shows from step 1. Default: FALSE}

\item{time_limit}{Number of days to include. Default = 90.}

\item{sim_population}{Size of population simulated. Only needed if providing 
`total_population`.}

\item{total_population}{True population size, needed only if simulation size 
is smaller than the true population size due to computational cost 
etc.}
}
\value{
\itemize{
\item \code{plot:} A \code{ggplot} object, bar charts of count of patients 
             requiring hospitalization and ICU respectively
\item \code{result:} A dataframe
   \itemize{\item \code{week:}  week number from input \code{sim},
            \item \code{hosp:} the number of patients that require hospitalization locally,
            \item \code{icu:} the number of patients that require ICU locally. }
}
}
\description{
Extract and plot information of local and weekly estimates from simulation
}
