% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD_infection.R
\name{infection.FUN}
\alias{infection.FUN}
\title{Infection function}
\usage{
infection.FUN(dat, at, seed = NULL)
}
\arguments{
\item{dat}{Merged input parameters.}

\item{at}{Step number}

\item{seed}{random seed}
}
\value{
Updated dat
}
\description{
Function to implement infection processes for SEIQHRF model
}
