% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD_init.R
\name{init_status.icm}
\alias{init_status.icm}
\title{Initialize status of ICM}
\usage{
init_status.icm(dat)
}
\arguments{
\item{dat}{Object containing all data}
}
\value{
Updated dat
}
\description{
Function to get the status of the initialized icm
}
