% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD_init.R
\name{initialize.FUN}
\alias{initialize.FUN}
\title{Initialize ICM}
\usage{
initialize.FUN(param, init, control, seed = NULL)
}
\arguments{
\item{param}{ICM parameters.}

\item{init}{Initial value parameters.}

\item{control}{Control parameters}

\item{seed}{random seed for checking consistency with other versions.}
}
\value{
Updated dat
}
\description{
Function to initialize the icm
}
