% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_seiqhrf.R
\name{merge.seiqhrf.icm}
\alias{merge.seiqhrf.icm}
\title{Merge icm}
\usage{
\method{merge}{seiqhrf.icm}(x, y, ...)
}
\arguments{
\item{x}{First to merge}

\item{y}{Second to merge}

\item{...}{other parameters}
}
\value{
Combined
}
\description{
Function to merge icms
}
