% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{plot.list}
\alias{plot.list}
\title{Plot simulation result given list}
\usage{
\method{plot}{list}(
  x,
  comp_remove = "none",
  time_limit = 90,
  ci = TRUE,
  sep_compartments = FALSE,
  trans = FALSE,
  known = NULL,
  start_date = ymd("2020-03-21"),
  x_axis = "Date (MM-DD)",
  plot_title = "",
  sim_population = 1000,
  total_population = NULL,
  ...
)
}
\arguments{
\item{x}{A list of seiqhrf objects returned from \code{\link{seiqhrf}}.}

\item{comp_remove}{Compartments to remove. Suggest c(s.num, r.num)}

\item{time_limit}{Number of steps (days) to plot.}

\item{ci}{T/F to include 95\% confidence intervals in sirplus plot.}

\item{sep_compartments}{T/F use faceting to show each compartment in a 
separate plot, only works if plotting a single simulation.}

\item{trans}{Y-axis transformation (e.g. log2, log10).}

\item{known}{Dataframe with known compartment numbers to plot alongside
projections}

\item{start_date}{Date for day 0.}

\item{x_axis}{Title for x-axis.}

\item{plot_title}{Title for whole plot. 
total_population}

\item{sim_population}{Size of population simulated. Only needed if providing 
`total_population`.}

\item{total_population}{True population size, needed only if simulation size 
is smaller than the true population size due to computational cost 
etc.}

\item{...}{Additional parameters}
}
\description{
Plot simulation result given list
}
