% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_prior_selector.R
\name{plot.prior}
\alias{plot.prior}
\title{Plot theoretical probability density of specified priors.}
\usage{
\method{plot}{prior}(x, at = NULL, ...)
}
\arguments{
\item{x}{A prior object returned by \code{\link{select_prior}}.}

\item{at}{A vector of time points to plot when an element in x has length greater than 1.}

\item{...}{Additional parameters.}
}
\description{
Plot theoretical probability density of specified priors.
}
