% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/original_plot_models.R
\name{plot_models}
\alias{plot_models}
\title{Plot models}
\usage{
plot_models(
  sims = baseline_sim,
  sim_id = "baseline",
  comp_remove = "none",
  time_lim = 100,
  trans = "na",
  known = NULL,
  start_date = ymd("2020-03-21"),
  x_axis = "Days since beginning of epidemic",
  plot_title = "SEIQHRF"
)
}
\arguments{
\item{sims}{Single or list of sims to plot}

\item{sim_id}{String or list of strings to name each facet}

\item{comp_remove}{Compartments to remove. Suggest c(s.num, r.num)}

\item{time_lim}{Number of steps (days) to plot.}

\item{trans}{Y-axis transformation (e.g. log2, log10). Default = none.}

\item{known}{Dataframe with known compartment numbers to plot alongside
projections}

\item{start_date}{Date for day 0. Default: ymd("2020-03-21"),}

\item{x_axis}{Title for x-axis. Default: 'Days since beginning of epidemic'}

\item{plot_title}{Title for whole plot. Default: 'SEIQHRF plot'}
}
\value{
ggplot2 object
}
\description{
Function to plot individuals models or multiple models for comparison.
}
