% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{plot_sirplus}
\alias{plot_sirplus}
\title{Wrapper for primary sirplus plotting function}
\usage{
plot_sirplus(
  x,
  comp_remove,
  time_limit,
  ci,
  sep_compartments,
  trans,
  known,
  start_date,
  x_axis,
  plot_title,
  sim_population,
  total_population,
  ...
)
}
\arguments{
\item{x}{A seiqhrf object (or list of multiple seiqhrf objects) returned 
from \code{\link{seiqhrf}}.}

\item{comp_remove}{Compartments to remove. Suggest c(s.num, r.num)}

\item{time_limit}{Number of steps (days) to plot.}

\item{ci}{T/F to include 95\% confidence intervals in sirplus plot.}

\item{sep_compartments}{T/F use faceting to show each compartment in a 
separate plot, only works if plotting a single simulation.}

\item{trans}{Y-axis transformation (e.g. log2, log10).}

\item{known}{Dataframe with known compartment numbers to plot alongside
projections}

\item{start_date}{Date for day 0.}

\item{x_axis}{Title for x-axis.}

\item{plot_title}{Title for whole plot.}

\item{sim_population}{Size of population simulated. Only needed if providing 
`total_population`.}

\item{total_population}{True population size, needed only if simulation size 
is smaller than the true population size due to computational cost}

\item{...}{Additional parameters}
}
\value{
ggplot2 object
}
\description{
Flexible function to generate sirplus plots (i.e. compartment counts over 
time). This function allows for plotting multiple experiments, viewing the 
plots of different scales (e.g. log2), plotting compartments separately,
adding 95\% CIs, and plotting known data along side the simulations.
}
