% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_plot.R
\name{plot_times}
\alias{plot_times}
\title{Plot compartment duration distributions}
\usage{
plot_times(sim)
}
\arguments{
\item{sim}{An seiqhrf object returned from function \code{\link{seiqhrf}}.}
}
\value{
ggplot2 object
}
\description{
Function to plot Duration frequency distributions. If multiple simulations 
were performed (nsim >1), durations from sims are appended to each other.
}
