% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_prior_selector.R
\name{print.prior}
\alias{print.prior}
\title{Print important information from specified prior.}
\usage{
\method{print}{prior}(x, ...)
}
\arguments{
\item{x}{A prior object returned by \code{\link{select_prior}}.}

\item{...}{Additional parameters.}
}
\description{
Print important information from specified prior.
}
