% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD_recovery.R
\name{recovery.FUN}
\alias{recovery.FUN}
\title{Progress icm}
\usage{
recovery.FUN(dat, at, seed = NULL)
}
\arguments{
\item{dat}{Object containing all data}

\item{at}{time point}

\item{seed}{random seed for checking consistency}
}
\value{
progress
}
\description{
Function to get progress of icms
}
