% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_saveout.R
\name{saveout_seiqhrf}
\alias{saveout_seiqhrf}
\title{Save icm}
\usage{
saveout_seiqhrf(dat, s, out = NULL)
}
\arguments{
\item{dat}{Merged input parameters.}

\item{s}{Step number or time point.}

\item{out}{initial out. Default = NULL}
}
\value{
list of output in the same format as \code{icm} object in \code{EpiModel}.
}
\description{
Function to organize simulation outputs
}
