% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CORE_seiqhrf.R
\name{seiqhrf}
\alias{seiqhrf}
\title{SEIQHRF function}
\usage{
seiqhrf(init, control, param, priors = NULL, seed = NULL)
}
\arguments{
\item{init}{Initial state values, returned by \code{\link{init_seiqhrf}}.}

\item{control}{Control parameters of the model, returned by \code{\link{control_seiqhrf}}.}

\item{param}{Model parameters, returned by \code{\link{param_seiqhrf}}.}

\item{priors}{A prior object returned by \code{\link{select_prior}}.}

\item{seed}{random seed for checking consistency}
}
\value{
Updated dat
}
\description{
Function to implement infection processes for SEIQHRF model
}
