% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PREP_prior_selector.R
\name{select_prior}
\alias{select_prior}
\title{Define prior distributions to stochastic parameters}
\usage{
select_prior(
  param,
  pars,
  prior.dist = NULL,
  prior.pars = list(),
  shift_only = FALSE
)
}
\arguments{
\item{param}{Model parameters, returned by \code{\link{param_seiqhrf}}.}

\item{pars}{A vector of names of random parameters.}

\item{prior.dist}{Probability prior distributions of the parameters in \code{par}.}

\item{prior.pars}{Distribution parameters of the probability distributions in \code{prior.dist}.}

\item{shift_only}{In effect only for parameters (in \code{param}) that have length longer than 1. 
If TRUE, only parameter used at the first time point is simulated from prior distribution, 
those used for the latter time points are shifted from the input values in param by the same amount as the first time point.}
}
\description{
Define prior distributions to stochastic parameters
}
