% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIN_summary.R
\name{summary.seiqhrf}
\alias{summary.seiqhrf}
\title{summary function}
\usage{
\method{summary}{seiqhrf}(object, ...)
}
\arguments{
\item{object}{seiqhrf object, returned by \code{\link{seiqhrf}}.}

\item{...}{Additional parameters.}
}
\value{
A list of compartments, each compartment is a list of three: 
\itemize{
\item \code{mean}: (t-dimensional vector) Mean of simulated counts of the compartment;
\item \code{sd}: (t-dimensional vector) Standard deviation of simulated counts of the compartment;
\item \code{CI}: (t by 2 matrix) 95\% confidence intervals of simulated counts of the compartment (assumed Gaussian),
\item \code{qntCI}: (t by 2 matrix) 95\% quantile confidence intervals,
}
where t is the total number of time points in the simulation (i.e. object$nstep).
}
\description{
Function to extract mean and standard deviation of all compartments 
for SEIQHRF model
}
