% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_experiments.R
\name{vary_param}
\alias{vary_param}
\title{Generating time-dependent parameters}
\usage{
vary_param(nstep, vals, timing)
}
\arguments{
\item{nstep}{Number of time steps to generate parameter values for.}

\item{vals}{List of parameter values to include over the nsteps.}

\item{timing}{List of the step numbers at which to start changes, with the
last number reflecting when to hit the last parameter value in vals.}
}
\value{
list of parameter values for length t
}
\description{
Function to generate parameter values for the time range of the simulation
that can change over time.
}
