% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD_departures.R
\name{departures.FUN}
\alias{departures.FUN}
\title{Departures function}
\usage{
departures.FUN(dat, at, seed = NULL)
}
\arguments{
\item{dat}{Merged input parameters.}

\item{at}{Step number}

\item{seed}{random seed for checking consistency}
}
\value{
Updated dat
}
\description{
Function to handle background demographics for the SEIQHRF model.
Specifically departures (deaths not due to the virus, and emigration).
}
